+++ stdlib/Markdown/src/Julia/Julia.jl
@@ -7,8 +7,14 @@
 
 include("interp.jl")
 
-@flavor julia [blocktex, blockinterp, hashheader, list, indentcode, fencedcode,
-               blockquote, admonition, footnote, github_table, horizontalrule, setextheader, paragraph,
+@flavor julia [blocktex, blockinterp,
+               list, indentcode, blockquote, admonition, footnote, hashheader,
+               fencedcode, github_table, horizontalrule, setextheader, paragraph,
 
-               linebreak, escapes, tex, interp, en_or_em_dash, inline_code,
-               asterisk_bold, underscore_bold, asterisk_italic, underscore_italic, image, footnote_link, link, autolink]
+               linebreak, escapes,
+               tex, interp,
+               en_or_em_dash, inline_code,
+               double_tilde_strikethrough, tilde_strikethrough,
+               asterisk_bold, underscore_bold,
+               asterisk_italic, underscore_italic,
+               image, footnote_link, link, autolink]